+++ app/src/proto/team_pb_service.js
@@ -1,0 +1,3 @@
+// package: team
+// file: team.proto
+