+++ app/src/utils/constants.ts
@@ -11,9 +11,15 @@ export const applicationsDescription = 'Applications are a Custom Resource for k
 // and the resources page.
 export const resourcesDescription = 'Resources are all available Kubernetes resources.';
 
+// teamsDescription is the description, which is displayed in the UI for the Team CR,
+export const teamsDescription = 'Teams are a Custom Resource for kobs.';
+
 // applicationAnnotation is the key for the annotation, which allows a user to specify a list of applications within a
 // resource
 export const applicationAnnotation = 'kobs.io/applications';
+
+// teamAnnotation is the key for the annotation, which allows a user to specify a list of teams within a resource.
+export const teamAnnotation = 'kobs.io/teams';
 
 // pluginAnnotation is the key for the annotation, which allows a user to specify a list of plugins within a resource.
 export const pluginAnnotation = 'kobs.io/plugins';